Reorganize package for simpler `go install`
